algo-machine-v3-github/algo-v3/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pandas==2.2.2
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.1
ta==0.11.0
scipy==1.13.0
aiofiles==23.2.1
